+++ packages/ui-core/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+- **Button**
+    - depricated few values of properties passiveColor and disabledColor
+    - add new values for propperties passiveColor and disabledColor
 
 ## Release 1.0.10
 - **ContentArea**